# Event.Namespaces — Faster iteration.