… dist/es6.observable.js
… dist/observable.js
… dist/riot.observable.js